BKNDLSS-24237 - Adapt dev profile to new routes and responses
